Minor refactor to decouple config from the rest of the app
