+++ src/config.rs
@@ -13,7 +13,6 @@ use toml;
 
 use crate::errors::Error;
 use crate::tls::{load_certs, load_keys};
-use crate::{ClientServerMap, ConnectionPool};
 
 pub const VERSION: &str = env!("CARGO_PKG_VERSION");
 
@@ -523,28 +522,6 @@ pub async fn parse(path: &str) -> Result<(), Error> {
     CONFIG.store(Arc::new(config.clone()));
 
     Ok(())
-}
-
-pub async fn reload_config(client_server_map: ClientServerMap) -> Result<bool, Error> {
-    let old_config = get_config();
-    match parse(&old_config.path).await {
-        Ok(()) => (),
-        Err(err) => {
-            error!("Config reload error: {:?}", err);
-            return Err(Error::BadConfig);
-        }
-    };
-    let new_config = get_config();
-
-    if old_config.pools != new_config.pools {
-        info!("Pool configuration changed, re-creating server pools");
-        ConnectionPool::from_config(client_server_map).await?;
-        Ok(true)
-    } else if old_config != new_config {
-        Ok(true)
-    } else {
-        Ok(false)
-    }
 }
 
 #[cfg(test)]
